Add next Meetup feature and update UI
This commit adds a new method to get the next Meetup in the `MeetupRepository`. The UI has also been updated in various places including the home page and base template to better present information about the next Meetup. Moreover, colors and new brand image are added in the `tailwind.config.js`.