Integrate MeetupRepository into Admin Dashboard
The DashboardController constructor now includes the MeetupRepository. It utilizes this repository to fetch the next Meetup event details for display in the admin dashboard. The admin index view is also updated to display the said details including a countdown to the event. Several modifications were also made to menu items and page titles.